css/css-values/minmax-length-serialize.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-values-4/#comp-func">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#lengths">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#calc-serialize">
<link rel="author" title="Xiaocheng Hu" href="mailto:xiaochengh@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/parsing-testcommon.js"></script>
<script>
function test_computed(value, expected) {
    test_valid_value('left', value, expected);
}

test_computed('min(1px)', 'calc(1px)');
test_computed('min(1in)', 'calc(96px)');
test_computed('max(1px)', 'calc(1px)');
test_computed('max(1in)', 'calc(96px)');

// Values are case-insensitive and serialize as lower case, for example 1Q
// serializes as 1q.
test_computed('min(1PX)', 'calc(1px)');

// Arguments simplify down eagerly
test_computed('min(50px, 1in + 1px)', 'calc(50px)');
test_computed('max(50px, 1in + 1px)', 'calc(97px)');

// And the entire function simplifies eagerly if possible
test_computed('calc(1px + min(1in, 100px)', 'calc(97px)');
test_computed('calc(1px + max(1in, 100px)', 'calc(101px)');

// Can't test that min()/max() are preserved in computed values with just lengths;
// see minmax-length-percentage-serialize for tests of that.
</script>
